requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.36.2
botocore==1.36.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Django==4.2.18
docker==7.1.0
docstring_parser==0.16
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.6
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
groq==0.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jmespath==1.0.1
kubernetes==31.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy==1.14.1
mypy-extensions==1.0.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.59.8
packaging==24.2
pgvector==0.3.6
phidata==2.7.8
pluggy==1.5.0
psycopg==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
ruff==0.9.2
s3transfer==0.11.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
starlette==0.41.3
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
websocket-client==1.8.0